+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- When using `check-deps`, do not throw exceptions that occur before the DotNetConsoleBuilder is built
+
 ## [5.0.0] - 2025-01-27
 
 ### Added